Refactor: 라우팅을 위한 폴더 디렉토리 구조 변경
