+++ src/app/(with-navbar)/about/_component/WorshipSchedule.tsx
