+++ src/app/(with-navbar)/about/page.module.scss
@@ -49,7 +49,7 @@
     left: 0;
     right: 0;
     bottom: 0;
-    background-image: url('../../../public/images/aboutBanner.jpg'); // 배경 이미지 경로
+    background-image: url('../../../../public/images/aboutBanner.jpg'); // 배경 이미지 경로
     background-size: 100% 100%;
     z-index: -1;
