+++ src/app/(with-navbar)/layout.tsx
@@ -1,0 +1,10 @@
+import { LayoutProps } from '@/shared/types/props';
+
+export default function NavbarLayout({ children }: LayoutProps) {
+  return (
+    <>
+      <nav>NavbarLayout</nav>
+      {children}
+    </>
+  );
+}
