+++ src/app/(with-navbar)/news/bulletin/[id]/page.tsx
