+++ src/app/(with-navbar)/news/bulletin/_component/BulletinTable.tsx
