+++ src/app/(with-navbar)/news/bulletin/create/page.tsx
