+++ src/app/(with-navbar)/news/bulletin/page.module.scss
