+++ src/app/(with-navbar)/news/page.tsx
