+++ src/app/(without-navbar)/gallery/[id]/page.tsx
