+++ src/app/(without-navbar)/gallery/page.tsx
