+++ src/app/(without-navbar)/layout.tsx
@@ -1,0 +1,5 @@
+import { LayoutProps } from '@/shared/types/props';
+
+export default function WithoutNavbarLayout({ children }: LayoutProps) {
+  return <>{children}</>;
+}
