+++ src/app/_component/home/Banner.tsx
