+++ src/app/_component/home/HomeGalleryCarousel.tsx
