+++ src/app/_component/home/HomeGalleryClient.tsx
