+++ src/app/_component/home/index.ts
@@ -1,9 +1,7 @@
 import Banner from './Banner';
-import Footer from './Footer';
-import Header from './Header';
 import HomeGallery from './HomeGallery';
 import GalleryCarousel from './HomeGalleryCarousel';
 import Greeting from './Greeting';
 import LinkButtons from './LinkButtons';
 
-export { Banner, Footer, Header, HomeGallery, GalleryCarousel, Greeting, LinkButtons };
+export { Banner, HomeGallery, GalleryCarousel, Greeting, LinkButtons };
