+++ src/app/_component/layout/Header.tsx
@@ -4,10 +4,10 @@ import { useEffect } from 'react';
 import Link from 'next/link';
 import { usePathname } from 'next/navigation';
 import { GiHamburgerMenu } from 'react-icons/gi';
-import Drawer from './Drawer';
-import ModalOverlay from './common/ModalOverlay';
-import UserProfile from './user/UserProfile';
-import UserProfileModal from './user/UserProfileModal';
+import Drawer from '../home/Drawer';
+import ModalOverlay from '../common/ModalOverlay';
+import UserProfile from '../user/UserProfile';
+import UserProfileModal from '../user/UserProfileModal';
 import { useProfile } from '@/context/SessionContextProvider';
 import useModal from '@/hooks/useModal';
 import useIsMobile from '@/hooks/useIsMobile';
