+++ src/shared/types/props.ts
@@ -1,0 +1,4 @@
+export type LayoutProps = {
+  children?: React.ReactNode;
+  params?: Promise<{ [key: string]: string | string[] | null }>;
+};
